+++ src/models/parameters.ts
@@ -177,6 +177,17 @@ export const expiryTime: OperationParameter = {
   }
 };
 
+export const regionBias: OperationParameter = {
+  parameterPath: ["options", "regionBias"],
+  mapper: {
+    serializedName: "regionBias",
+    xmlName: "regionBias",
+    type: {
+      name: "String"
+    }
+  }
+};
+
 export const accept1: OperationParameter = {
   parameterPath: "accept",
   mapper: {