+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow custom resume fields to be nullable
 - Allow custom job description fields to be nullable
 - Make "pdf" property in SearchResults nullable
+- Nest line item table rows correctly.
 
 ### Fixed
 - Allow rejectDuplicates to be null
@@ -25,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add custom data to job description search results
 - Add international_country_code to phone number details
 - Add customData field to JD search param/result serializers
+- rawText field to JobDescription Model
 
 ## [5.3.1] - 2023-03-30
 ### Changed